refactor: useConfirm과 useAlert를 useDialog로 통합
