+++ src/hooks/dialog/useDialog.tsx
@@ -1,7 +1,7 @@
 import { useOverlay } from "@hooks/use-overlay/useOverlay.tsx";
 import Dialog from "@hooks/dialog/Dialog.tsx";
 
-export const useConfirm = () => {
+export const useDialog = () => {
   const { openOverlay, closeOverlay } = useOverlay();
   const openConfirm = ({
     title,
@@ -34,5 +34,30 @@ export const useConfirm = () => {
     });
   };
 
-  return { openConfirm };
+  const openAlert = ({
+    title,
+    content,
+    approveText,
+  }: {
+    title: string;
+    content: string;
+    approveText: string;
+  }): Promise<boolean> => {
+    return new Promise((resolve) => {
+      openOverlay(
+        <Dialog
+          title={title}
+          content={content}
+          approveText={approveText ?? undefined}
+          onClickApprove={() => {
+            resolve(true);
+            closeOverlay();
+          }}
+          onClickReject={() => {}}
+        />
+      );
+    });
+  };
+
+  return { openConfirm, openAlert };
 };
